requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
Pillow==7.2.0
pony==0.7.13
psycopg2==2.8.6
psycopg2-binary==2.8.6
requests==2.24.0
six==1.15.0
urllib3==1.25.10
vk-api==11.9.0
